Add check for `Sprint` field
Checks for current sprint based on start date and duration of iteration, reporting it in a custom field named `{IterationName}/IsCurrent` with Y/N.

`main_test` might have to be updated later as it will probably fail when a PR is made outside of current sprint period. This may not be necessary, depending on implementation of another issue.